+++ editor/editor_audio_buses.cpp
@@ -1440,7 +1440,7 @@ Size2 EditorAudioMeterNotches::get_minimum_size() const {
 	float width = 0;
 	float height = top_padding + btm_padding;
 
-	for (uint8_t i = 0; i < notches.size(); i++) {
+	for (int i = 0; i < notches.size(); i++) {
 		if (notches[i].render_db_value) {
 			width = MAX(width, font->get_string_size(String::num(Math::abs(notches[i].db_value)) + "dB").x);
 			height += font_height;
@@ -1474,7 +1474,7 @@ void EditorAudioMeterNotches::_draw_audio_notches() {
 	Ref<Font> font = get_font("font", "Label");
 	float font_height = font->get_height();
 
-	for (uint8_t i = 0; i < notches.size(); i++) {
+	for (int i = 0; i < notches.size(); i++) {
 		AudioNotch n = notches[i];
 		draw_line(Vector2(0, (1.0f - n.relative_position) * (get_size().y - btm_padding - top_padding) + top_padding),
 				Vector2(line_length, (1.0f - n.relative_position) * (get_size().y - btm_padding - top_padding) + top_padding),
